Split up crypto/sha2
