+++ src/hash.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "hash.h"
+#include "crypto/hmac_sha512.h"
 
 inline uint32_t ROTL32(uint32_t x, int8_t r)
 {
