+++ src/hash.h
@@ -7,7 +7,7 @@
 #define BITCOIN_HASH_H
 
 #include "crypto/ripemd160.h"
-#include "crypto/sha2.h"
+#include "crypto/sha256.h"
 #include "serialize.h"
 #include "uint256.h"
 #include "version.h"
