+++ src/script/interpreter.cpp
@@ -8,7 +8,7 @@
 #include "core/transaction.h"
 #include "crypto/ripemd160.h"
 #include "crypto/sha1.h"
-#include "crypto/sha2.h"
+#include "crypto/sha256.h"
 #include "eccryptoverify.h"
 #include "pubkey.h"
 #include "script/script.h"
