+++ src/test/crypto_tests.cpp
@@ -4,7 +4,9 @@
 
 #include "crypto/ripemd160.h"
 #include "crypto/sha1.h"
-#include "crypto/sha2.h"
+#include "crypto/sha256.h"
+#include "crypto/sha512.h"
+#include "crypto/hmac_sha512.h"
 #include "random.h"
 #include "utilstrencodings.h"
